refactor: extract server spec deserializer and relocate auth restlet test
