+++ src/main/java/io/naftiko/spec/scripting/OperationStepScriptSpec.java
